+++ test/file_system_test.rb
@@ -38,6 +38,11 @@ describe Shrine::Storage::FileSystem do
       FileUtils.rmdir(root)
     end
 
+    it "expands the directory" do
+      assert_equal File.expand_path("test"),        file_system("test").directory.to_s
+      assert_equal File.expand_path("test/prefix"), file_system("test", prefix: "prefix").directory.to_s
+    end
+
     it "creates the given directory" do
       @storage = file_system(root)
       assert File.directory?(root)